Moved small implementation files to same files with API.
