+++ src/main/java/pl/bdygasinski/filewalker/ContentProvider.java
@@ -1,5 +1,7 @@
 package pl.bdygasinski.filewalker;
 
+import java.io.IOException;
+import java.io.UncheckedIOException;
 import java.nio.file.Path;
 import java.util.Set;
 
@@ -10,4 +12,18 @@ public interface ContentProvider {
     }
 
     Set<Entry> provideEntriesFrom(Path path);
+}
+
+class ConsoleContentProvider implements ContentProvider{
+
+    @Override
+    public Set<Entry> provideEntriesFrom(Path path) {
+        try {
+            Entry rootEntry = Entry.fromPathOrThrow(path);
+            return rootEntry.getVisibleRootLevelOrThrow();
+
+        } catch (IOException e) {
+            throw new UncheckedIOException(e);
+        }
+    }
 }
